Major update, almoost all done: + getnNumber now works with "-a" +! deleting all spaces at the begginging; next step will be do delete all stuff like: 	while(formula[i] == ' ') 		i++; + added step alternatng args :Z(from,to,step) + improved table alternating args + gAlternatingArgAmount now resets on each isFormulaCorrect ! need to handle opening and closing brackets in a better way (look for 'saved' in code) + check argument and getAlternateArgValues divided into separate function

+++ ONP/Functions.h
@@ -75,7 +75,7 @@ double factorialFunction(vector<double> args)
 		return 1;
 	else
 	{
-		int a = args[0];
+		int a = (int)args[0];
 		while(--a)
 			args[0] *= a;
 		return args[0];
